requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-experimental
sqlalchemy
streamlit
python-dotenv
    